+++ CHANGES.md
@@ -1,3 +1,13 @@
+## Version 1.2
+
+### Major changes
+
+- Added the (optional) generation of a JSON-LD `@context` file
+
+### Minor changes, bug fixes
+
+- Better use of the Commander package for the CLI
+
 ## Version 1.1
 
 - Restructured the HTML output: each term in its own subsection instead of into a table (this gives more space for the content)